# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Route & Schedule Service API",
        "description": "REST API documentation for Route & Schedule Service in IAE Assignment 2.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost",
            "description": "Local Docker server"
        }
    ],
    "paths": {
        "/api/v1/routes": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get all routes",
                "operationId": "6bcf29190fb0bd47d4e89c5fb951a4ca",
                "responses": {
                    "200": {
                        "description": "Routes retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                },
                "security": [
                    {
                        "IAEApiKey": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Routes"
                ],
                "summary": "Create new route",
                "operationId": "1b19d3d42e900f7c3886f64da1b52401",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "route_code",
                                    "origin",
                                    "destination",
                                    "departure_point",
                                    "arrival_point",
                                    "departure_date",
                                    "departure_time",
                                    "arrival_time",
                                    "vehicle_type",
                                    "price",
                                    "seat_capacity",
                                    "available_seats",
                                    "status"
                                ],
                                "properties": {
                                    "route_code": {
                                        "type": "string",
                                        "example": "BDG-JKT-003"
                                    },
                                    "origin": {
                                        "type": "string",
                                        "example": "Bandung"
                                    },
                                    "destination": {
                                        "type": "string",
                                        "example": "Jakarta"
                                    },
                                    "departure_point": {
                                        "type": "string",
                                        "example": "Pool Dago"
                                    },
                                    "arrival_point": {
                                        "type": "string",
                                        "example": "Terminal Lebak Bulus"
                                    },
                                    "departure_date": {
                                        "type": "string",
                                        "example": "2026-05-23"
                                    },
                                    "departure_time": {
                                        "type": "string",
                                        "example": "09:00"
                                    },
                                    "arrival_time": {
                                        "type": "string",
                                        "example": "12:30"
                                    },
                                    "vehicle_type": {
                                        "type": "string",
                                        "example": "Travel Executive"
                                    },
                                    "price": {
                                        "type": "integer",
                                        "example": 150000
                                    },
                                    "seat_capacity": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "available_seats": {
                                        "type": "integer",
                                        "example": 12
                                    },
                                    "status": {
                                        "type": "string",
                                        "example": "available"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Route created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "422": {
                        "description": "Validation failed"
                    }
                },
                "security": [
                    {
                        "IAEApiKey": []
                    }
                ]
            }
        },
        "/api/v1/routes/{id}": {
            "get": {
                "tags": [
                    "Routes"
                ],
                "summary": "Get route by ID",
                "operationId": "488caa4ed6b252e87fa96f956e099b99",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Route retrieved successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Route not found"
                    }
                },
                "security": [
                    {
                        "IAEApiKey": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "IAEApiKey": {
                "type": "apiKey",
                "name": "X-IAE-KEY",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Routes",
            "description": "Routes"
        }
    ]
}